Made Suggested Changes
