+++ plinth/modules/pagekite/forms.py
@@ -27,7 +27,6 @@ import logging
 
 from plinth import cfg
 from plinth.errors import ActionError, DomainRegistrationError
-from plinth.modules.first_boot.forms import SubdomainWidget
 from plinth.modules.pagekite.utils import PREDEFINED_SERVICES, run
 from plinth.utils import format_lazy
 
@@ -38,12 +37,30 @@ LOGGER = logging.getLogger(__name__)
 
 class TrimmedCharField(forms.CharField):
     """Trim the contents of a CharField"""
+
     def clean(self, value):
         """Clean and validate the field value"""
         if value:
             value = value.strip()
 
         return super(TrimmedCharField, self).clean(value)
+
+
+class SubdomainWidget(forms.widgets.TextInput):
+    """Append the domain to the subdomain bootstrap input field"""
+
+    def __init__(self, domain, *args, **kwargs):
+        """Intialize the widget by storing the domain value."""
+        super().__init__(*args, **kwargs)
+        self.domain = domain
+
+    def render(self, *args, **kwargs):
+        """Return the HTML for the widget."""
+        inputfield = super().render(*args, **kwargs)
+        return """<div class="input-group">
+                  {0}
+                  <span class="input-group-addon">{1}</span>
+               </div>""".format(inputfield, self.domain)
 
 
 class ConfigurationForm(forms.Form):
@@ -54,9 +71,9 @@ class ConfigurationForm(forms.Form):
 
     server_domain = forms.CharField(
         label=ugettext_lazy('Server domain'), required=False,
-        help_text=\
-        ugettext_lazy('Select your pagekite server. Set "pagekite.net" to use '
-                      'the default pagekite.net server.'),
+        help_text= \
+            ugettext_lazy('Select your pagekite server. Set "pagekite.net" to use '
+                          'the default pagekite.net server.'),
         widget=forms.TextInput())
     server_port = forms.IntegerField(
         label=ugettext_lazy('Server port'), required=False,
@@ -70,9 +87,9 @@ class ConfigurationForm(forms.Form):
 
     kite_secret = TrimmedCharField(
         label=ugettext_lazy('Kite secret'),
-        help_text=\
-        ugettext_lazy('A secret associated with the kite or the default secret '
-                      'for your account if no secret is set on the kite.'))
+        help_text= \
+            ugettext_lazy('A secret associated with the kite or the default secret '
+                          'for your account if no secret is set on the kite.'))
 
     def save(self, request):
         """Save the form on submission after validation."""
@@ -84,14 +101,14 @@ class ConfigurationForm(forms.Form):
             config_changed = False
 
             if old['kite_name'] != new['kite_name'] or \
-                    old['kite_secret'] != new['kite_secret']:
+                            old['kite_secret'] != new['kite_secret']:
                 utils.run(['set-kite', '--kite-name', new['kite_name']],
                           input=new['kite_secret'].encode())
                 messages.success(request, _('Kite details set'))
                 config_changed = True
 
             if old['server_domain'] != new['server_domain'] or \
-                    old['server_port'] != new['server_port']:
+                            old['server_port'] != new['server_port']:
                 server = "%s:%s" % (new['server_domain'], new['server_port'])
                 utils.run(['set-frontend', server])
                 messages.success(request, _('Pagekite server set'))
@@ -191,7 +208,6 @@ class BaseCustomServiceForm(forms.Form):
 
 
 class DeleteCustomServiceForm(BaseCustomServiceForm):
-
     def delete(self, request):
         service = self.convert_formdata_to_service(self.cleaned_data)
         utils.run(['remove-service', '--service', json.dumps(service)])
@@ -244,6 +260,8 @@ class AddCustomServiceForm(BaseCustomServiceForm):
                 messages.error(request, _('This service already exists'))
             else:
                 raise
+
+
 class State5Form(forms.Form):
     """Set up freedombox.me pagekite subdomain"""
     DOMAIN_APPENDIX = '.freedombox.me'
