+++ plinth/modules/users/forms.py
@@ -28,8 +28,9 @@ from django.utils.translation import ugettext as _, ugettext_lazy
 from plinth import actions
 from plinth.errors import ActionError
 
-# Usernames used by optional services (that might not be installed yet).
 from plinth.modules.security import set_restricted_access
+
+# Usernames used by optional services (that might not be installed yet).
 
 RESERVED_USERNAMES = [
     'debian-deluged',
@@ -55,11 +56,12 @@ GROUP_CHOICES = (
 
 class ValidNewUsernameCheckMixin(object):
     """Mixin to check if a username is valid for created new user."""
+
     def clean_username(self):
         """Check for username collisions with system users."""
         username = self.cleaned_data['username']
         if self.instance.username != username and \
-           not self.is_valid_new_username():
+                not self.is_valid_new_username():
             raise ValidationError(_('Username is taken or is reserved.'),
                                   code='invalid')
 
@@ -93,14 +95,14 @@ class CreateUserForm(ValidNewUsernameCheckMixin, UserCreationForm):
         label=ugettext_lazy('Groups'),
         required=False,
         widget=forms.CheckboxSelectMultiple,
-        help_text=\
-        ugettext_lazy('Select which services should be available to the new '
-                      'user. The user will be able to log in to services that '
-                      'support single sign-on through LDAP, if they are in the '
-                      'appropriate group.<br /><br />Users in the admin group '
-                      'will be able to log in to all services. They can also '
-                      'log in to the system through SSH and have '
-                      'administrative privileges (sudo).'))
+        help_text= \
+            ugettext_lazy('Select which services should be available to the new '
+                          'user. The user will be able to log in to services that '
+                          'support single sign-on through LDAP, if they are in the '
+                          'appropriate group.<br /><br />Users in the admin group '
+                          'will be able to log in to all services. They can also '
+                          'log in to the system through SSH and have '
+                          'administrative privileges (sudo).'))
 
     def __init__(self, request, *args, **kwargs):
         """Initialize the form with extra request argument."""
@@ -130,7 +132,7 @@ class CreateUserForm(ValidNewUsernameCheckMixin, UserCreationForm):
                     messages.error(
                         self.request,
                         _('Failed to add new user to {group} group.')
-                        .format(group=group))
+                            .format(group=group))
 
                 group_object, created = Group.objects.get_or_create(name=group)
                 group_object.user_set.add(user)
@@ -144,12 +146,12 @@ class UserUpdateForm(ValidNewUsernameCheckMixin, forms.ModelForm):
         label=ugettext_lazy('SSH Keys'),
         required=False,
         widget=forms.Textarea,
-        help_text=\
-        ugettext_lazy('Setting an SSH public key will allow this user to '
-                      'securely log in to the system without using a '
-                      'password. You may enter multiple keys, one on each '
-                      'line. Blank lines and lines starting with # will be '
-                      'ignored.'))
+        help_text= \
+            ugettext_lazy('Setting an SSH public key will allow this user to '
+                          'securely log in to the system without using a '
+                          'password. You may enter multiple keys, one on each '
+                          'line. Blank lines and lines starting with # will be '
+                          'ignored.'))
 
     class Meta:
         """Metadata to control automatic form building."""
@@ -244,8 +246,10 @@ class UserChangePasswordForm(SetPasswordForm):
 
         return user
 
+
 class State1Form(ValidNewUsernameCheckMixin, auth.forms.UserCreationForm):
     """Firstboot state 1: create a new user."""
+
     def __init__(self, *args, **kwargs):
         self.request = kwargs.pop('request')
         super().__init__(*args, **kwargs)
